Update PerformanceLevelForm to use FloatField for points and align template

+++ forms.py
@@ -409,7 +409,7 @@ class PerformanceLevelForm(FlaskForm):
     min_marks = IntegerField('Minimum Mark', validators=[DataRequired(), NumberRange(min=0, max=100)])
     max_marks = IntegerField('Maximum Mark', validators=[DataRequired(), NumberRange(min=0, max=100)])
     level = StringField('Level', validators=[DataRequired(), Length(min=1, max=50)])
-    points = IntegerField('Points', validators=[DataRequired(), NumberRange(min=0, max=100)])
+    points = FloatField('Points', validators=[DataRequired(), NumberRange(min=0, max=100)])
     comment = StringField('Comment', validators=[DataRequired(), Length(min=1, max=200)])
     type = SelectField('Type', choices=[
         ('learning_area', 'Learning Area'),
